todo el back corregido y front comenzado. Ya fue aceptado por cliente y se deberán agregar detalles

+++ backend/src/main/java/una/data/ConfiguracionEventoRepository.java
@@ -3,5 +3,6 @@ package una.data;
 import una.logic.ConfiguracionEvento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConfiguracionEventoRepository extends JpaRepository<ConfiguracionEvento, Integer> {
+public interface ConfiguracionEventoRepository extends JpaRepository<ConfiguracionEvento, Byte> {
+    // el segundo parámetro genérico cambia de Integer a Byte
 }
